MSVC: don't wait for wineserver to exit
`wineserver -w` waits for a global wineserver instance to exit, meaning that even though Buildmaster runs 14 jobs in parallel, MSVC jobs wait for *all* other running wine processes to finish before exiting. Some tests with wine on `buildpack-deps:bionic` suggest that `wine` returns when the process it's running terminates, so this may not be necessary.

Note that this is not tested with MSVC itself, which may behave differently.