Explicitly set CMAKE_AR
It's hard to find what this should be set to. Most examples online use msys or llvm. The one example I found that seems to use the MSVC toolchain is https://www.mail-archive.com/cmake@cmake.org/msg61272.html which indicates that the linker also serves as the archiver. I have no experience with the MSVC toolchain, though, and I have no idea whether this is accurate.